pyproject.toml
[project]
name = "manzai-archive-pipeline"
version = "0.1.0"
description = "Ingest video URLs and produce speaker-labeled manzai scripts"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "faster-whisper>=1.1.0",
    "pyannote.audio>=3.3.0",
    "torch>=2.4",
    "torchaudio>=2.4",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-slugify>=8.0",
    "python-dotenv>=1.0",
]

[project.scripts]
manzai = "pipeline.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]
